requirements.txt
PyYAML>=6.0
biopython>=1.81
semanticscholar>=0.5.0
pyzotero>=1.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0
